+++ Packages/com.quickeye.ui-toolkit-plus/Editor/QuickEye.UxmlBridgeGen/Model/MemberIdentifierStyle.cs
@@ -1,0 +1,16 @@
+using System;
+
+namespace QuickEye.UxmlBridgeGen
+{
+    [Serializable]
+    internal struct MemberIdentifierStyle
+    {
+        public string prefix;
+        public string suffix;
+        public CaseStyle style;
+        public string Apply(string identifier)
+        {
+            return $"{prefix}{identifier.ApplyCaseStyle(style)}{suffix}";
+        }
+    }
+}